Agent 6 Implementation
A working version of Agent 6. Perhaps a little slow (if we need to, we can do some optimizations), but otherwise, this should be a working version of Agent 6. This framework should be able to be easily applied to Agent 7, and perhaps it's even useful for much of Agent 8. Let me know if there are any issues or questions or concerns.